Make sure files are written to the correct dir
The URL download storage provider needed to be rewritten to drop the use
of resty as the library didn't provide a way to save arbitrary files
into a directory on a per-request basis. Attempting to call
`SetOutputDirectory` on the client would modify the client so that _all_
users of the client at that time would output to that directory, rather
than just for the current request.

Also adds tracing to URL downloads.

Fixes #1978
Part of #1925