Add tracing to the docker client
Trace calls to docker, so we can see how long Docker calls take. This
changes the `dockerutils` package into a 'proper' client and embeds the
traced client within it.

Spans for calls to Docker will only end when the call logically ends
such as when the call returns or when the `io.ReadCloser`, which would
be retrieving the contents asynchronously, is closed.

Part of #1925