Add the span, trace & job IDs to log statements
This ensures that log statements contain the job ID and tracing
information to make it easier to correlate traces with the log
statements associated with them.

This also updates as many log statements as possible to use `log.Ctx`
so that the log statements include the node, span, trace & job IDs.

Fixes #1152