Log all messages in the correct format
We need to avoid outputting logging messages before logging has been
properly configured, otherwise it may be in the wrong format - JSON
format vs human-readable. We also need to be careful so that, if
something goes wrong _before_ logging is configured, the messages are
still logged (although in an arbitrary format rather than one the user
selected). To avoid causing too much refactoring, we also need to _not_
buffer logs when running as a test.

This adds a simple buffer to store log messages so that can be printed
out once logging has been configured.

Fixes #2004