Fixes incorrect state using list command
When using the CLI `list` command, the state shown for some jobs gives
the impression that the job has failed.

If a particular shard reports mixed states, for example 2xBidRejected
and 1xBidAccepted, then the list command will report BidRejected even
though there is an execution happening on that shard. This happens
because the function generating the summary depends on the order/value
of the ExecutationStateType enumeration, where Accepted and Rejected are
for the purposes of this task in the wrong order.

Rather than re-order the enumeration which may be depended on elsewhere,
this PR will report Complete or Active states back to the list command
before it reports 'Failure' states.

Fixes #1993