Move SystemManager & span handling to root command
Moving the creation & clean up of the SystemManager and root span avoids
having the same code copied between all the commands. This also gives
the opportunity to give the root spans a more coherent name rather than
the code location.

This also fixes an issue with the wasm command which didn't apply the
root commands PersistentPreRun.

Fixes #2023