Add support for passing a context when cleaning up
This adds a second type of clean up function - one that takes a
`context.Context`. This makes it easier to avoid problems of using
a cancelled context when cleaning up, as the logic doesn't need to be
spread about, and makes it easy to migrate a portion of log statements
that aren't using `log.Ctx` to do so.

Part of #2001