Apply Job selection policies in the Requestor node (#1961)
Here we give the Requestor node the ability to filter the jobs that it
accepts, so that it can take on a "gateway" role.

We re-use the Bid Strategies from the compute node – the reason being
that we might soon want to share them e.g. access control should be
enforced at both requestor and compute node, for defence in depth, so
common code makes that quicker.

The Requestor node responds to the same flags that the compute
nodes do in `bacalhau serve` e.g. allow switching on networked jobs.