Adds a frontend to the cancel command.
Adds more information on what the cancellation command is currently
doing, so that the user can see the steps involved.

In addition this now checks the current state of the job before
submitting the response to ensure the job is not already terminated.

A new spinner is used to wrap the underlying library to simplify its
usage and make it easier to swap out at a later time.