Executors can supply a BidStrategy
Previously, all of our bid strategies have been applicable regardless
of what type of job is being run. However, there are some requirements
that are executor-specific, such as the Docker image platform matching
the engine platform.

This commit specifies that Executors must now return a Bid Strategy.
For all executors this is currently just an empty BidStrategy that
does nothing. But the infrastructure is now present to allow us to
define executor-specific strategies, which we will do in the next
commit.