Moves the telemetry DetachedContext to util
The DetachedContext is currently defined in telemetry, but has wider
potential use in long-running goroutines that we don't want to fail when
their parent is cancelled.  For this reason this PR moves it into the
util pkg so that it can be used for the notify* functions of the
scheduler.