Ensure long running jobs have own context
The scheduler has several notify* functions that spawn a new goroutine
before returning on the current thread of execution - currently each has
a context created from a parent that is a HTTP request, so all of the
goroutines created from that context are closed when the request is
finished.

To avoid this, each of the notify* functions that create a new
goroutine, do so with a DetachedContext which although it provides
access to the parent's values, it does not response to the parents
cancel() or timeout.

Although this will stop goroutines being killed when the request has
finished, it does mean that the original context is kept around for
lifetime of the new context, resulting in more garbage.