Removes cancelled task from logging sensor
Currently the ActiveJobs logging_sensor will continue to report a job,
even if it has been cancelled. The job is recorded as having the
previous state to the cancellation.

This PR removes the job from the map of running/active jobs so that it
does not continue to show as running.