Interpret job failures correctly in the CLI
At the moment the client does not correctly interpret overall job
failures. These are rare and in fact only really happen when we don't
receive enough bids but should also start happening when we retry
failed executions but run out of nodes.

This commit updates the client to understand job-level failures, and
correctly display an error response and stop polling for events.