Requester handles bid responses needing async approval
Previously job selection was a synchronous affair and happened within
the lifecycle of the job submission call. Now that we have responses
that can call for waiting, the requester needs to not start the job
if we are asked to wait.

There is now a job queue that wraps a scheduler. Jobs are enqueued by
default and only if the bid strategy returns a non-wait response is
the job immediately dequeued. Otherwise, the job can be manually
approved by a pre-configured client.