Allow HTTP job selectors to return a wait response
Previously our HTTP job selector only used the status code of the HTTP
request to determine whether or not we should bid on the job, and
ignored the content. This works but now that we are introducing wait
states for bid responses we need to send at least two bits of
information. It would also be nice to have a reason supplied!

Now, we allow HTTP responses to contain a full bid response object if
they also use an `application/json` content type. The response object
is returned directly if they do this.