Allow in-process IPFS to find a random port
Use the built-in multiaddress functionality to find a random port rather
than attempting to find a port before starting up IPFS and hoping the
port stays free. The API server is also changed to allow the port to be
set to 0 and so reduce the use of the `freeport` dependency.

Also fixes a flake in the `TestWebsocketSuite/TestWebsocketEverything`
test where the websocket connected _after_ the first job had been
submitted.

Fixes #2144