Implement a way to get output from a running job (#2159)
As we require a mechanism to capture and transport an execution's stdout
and stderr, this commit introduces a new method to the compute Executor
interface for executors to implement.

Initially we support only docker, with support for wasm coming at a
future time. In addition, we re-use the framing that docker uses when
multiplexing stdout and stderr across a single stream, something we can
re-use for wasm and transport across the network to a requester node.
The format of the data frame is documented in
pkg/util/logstream/dataframe.go.