Add tracing around WASM execution
While the Docker executor can rely on simply tracing the interactions
with the Docker client to produce a sensible trace that identifies the
key areas of work, no such tactic can be used with WASM as it all is
performed in-process. This adds spans around compiling modules,
instantiating modules, and calling functions.

Fixes #1990