enable retries with pluggable retry strategies (#2205)
Retry failed executions on other nodes in the network, including
failures to execute the job, verification failures, failure to publish
the results, and having the ask for bid rejected by the initially
selected nodes.

Retry strategy is pluggable, and multiple implementations can be
chained. Currently we only have a single strategy that always retries as
long as there are available nodes in the network to retry on. More
strategies can be implemented in the future, specially after we have
better breakdown if execution errors, and be able to differentiate
between bad jobs and retriable node failures

closes #1626