Simplify client interface for setting host and port
Previously our client(s) required specifying the base URL of the API
to configure what requester/compute node the client will talk to. This
was problematic for two reasons:

1. It requires specifying the protocol (as 'http') which doesn't make
   sense if our client calls actually use websockets or we upgrade to
   HTTPS in the future.
2. It requires all consumers to do a `fmt.Sprintf` as nearly always
   people have a host/port pair rather than a full URL.

Now the clients take hostname and port instead, so people can just
supply these directly and it is up to the client what protocol it
talks to the endpoint. We also use a proper uint16 to represent ports.