Automatically resolve WebAssembly module imports
Previously, WebAssembly module imports could be specified in the job
spec and these would be added to the WASM runtime before execution.
This required the user to specify all of the required modules,
including any dependencies of dependencies.

Now, the WASM executor can read module names as storage locations and
automatically download any required modules. The user just needs to
specify a suitably formatted module and all dependencies will be
loaded.

This also means we only instantiate WASI when we really need it.