Enforce lints in CI rather than locally
Enforcing linting locally is not having the desired effect because it
only checks that the hooks are installed when running the Makefile. It
is clear that many contributors are not even running make! Plus, the
git hooks fail in odd ways when run locally.

So instead we now don't require hooks to be install (but recommend
them still) and ensure that they all pass in CI.