Return error in execution failure for events.
Currently when there is an execution failure, such as a missing binary
in a docker image, the error is returned by the `describe` command as
the execution status.  Unfortunately, the command line relies on the
events to determine the current state of the job, and although these
contain a Comment field, the status is not present.

This PR sets the Comment on the event for an execution failure.