Allow fetching of events/history 'since' a specific time (#2238)
Currently in the CLI we fetch the event history of a job and filter it
client-side, as well as maintaining state (what have we seen and
processed). This PR adds a `since` parameter to the GetJobHistory method
of the store, allowing for the filtering to be done earlier (and thus
reducing the amount of records moved unncessarily). This also adds
support in the ApiClient, but does not provide a mechanism for the user
to specify this value in the CLI.