Enables filtering of job history
When retrieving the JobHistory from a requester node, all of the history
is provided, whether it is a job level event, or an execution event. The
history is also not ordered and often results in the caller retrieving
the entire history and filtering on the client side.

This PR introduces filter options to the job history, allowing for all
of the events after (or equal) to a specific epoch time, with one second
resolution. It is also possible to only retrieve job level events, or
only execution level. The results are also sorted according to their
`Time` field (with <1s resolution).