Only ask for livestream if following
Currently the reader always asks for the livestream when it gets to the
end of the file, even if it will then not read from the channel for
input.  By moving it into the Read() call and only when `follow` is set
this will stop the log manager writing anything to the unused channel.