Dashboard-based job moderation
This commit upgrades the dashboard to allow users to complete job
moderation requests.

A dashboard API can now be specified as an external HTTP hook at both
requester and compute nodes. When the dashboard receives the call, it
will store the request and respond with a wait request – so that the
job will be queued rather than executed.

Once a human has clicked the Moderate button in the dashboard UI, the
dashboard will call back to either the compute or requester node which
will schedule or cancel the job accordingly.

Currently, the dashboard returns a wait request in all circumstances,
i.e. it does not have any concept of “this job should be moderated and
this other job should not”. It also does not provide a list of open
requests yet so the user has to just find the job based on the ID. As
such, it’s not ready for production use yet.

This commit also introduces the first tests of the dashboard used in
tandem with a Bacalhau network!