Fixes race condition with spinner when retry fails (#2236)
As we now have retries, the behavious of the CLI has to change to
support it. We can no longer trust that the ExecutionStates previous
used will reflect the state of our submission - an execution failure no
longer means the CLI should report an error, instead we should rely on
the JobState changes instead.

Now the state mapping is 

```
JobStateNew        -> Creating job for submission
JobStateQueued     -> Job waiting to be scheduled
JobStateInProgress -> Job in progress
JobStateError      -> Error while executing the job
JobStateCancelled  -> Job canceled
JobStateCompleted  -> Job finished
```

At the same time, this PR switches the CLI over to the same spinner used
by the cancel command to simplify the processing of the job submission.