Allow compute nodes to receive async job authorizations
The compute node has always been able to call out to an external web
hook as part of the bidding process but until now it has not been
able to handle a waiting response and was required to respond to bids
immediately. The compute node therefore had no need for a web
callback once the job was moderated.

Now, the bid response is an asynchronous operation and the compute
node can receive its own web callback. Compute nodes acknowledge that
they have received a bid request but will respond to it in an async
manner.

This allows compute nodes to exist in a separate trust domain from the
requester node, because they can now have their own dashboard which
will moderate jobs which can be separate from anything else.