Removes the result folder immediately after publish
Currently, the results folders are queued up the verifiers to cleanup
the folders when the node is shut down. This means that until then there
are a lot of unnecessary folders on disk.

To resolve this the code now deletes the results folder immediately
after publishing it to ensure the space is freed up immediately.