Apply job selection policies in the dashboard
At the moment the dashboard will require all jobs to be approved by a
human. However, using the job selection policies, some jobs can be
automatically approved, and only the rest can be moderated.

This commit adds the same job selection flags used by the compute and
requester nodes to the dashboard. When the dashboard receives a job,
it will check the job against its configured policies and immediately
return a positive response if it can. Else, the job will be held for
moderation like it would be currently and can be approved or rejected
by a human.