Ensure the minimum number of bids is reached before execution starts
Due to a bug introduced in e866a9, the requester node was no longer
able to remember if it had received enough accepted bids and assumed
that any execution it knew about was accepted.

This commit restores the ability of the requester to remember bid
responses and adds a test that ensures the requester cannot start the
job until the min bid threshold has been reached, including that it
cancels the job if necessary.

Co-authored-by: frrist <forrest@protocol.ai>