Allows bprotocol to process errors with responses
Currently the compute node end of the bprotocol is unable to return
errors to the requester node, resulting in the requester node generating
errors about decoding the request object.

What we want to happen is to allow the Compute Proxy (and the Callback
Proxy) to receive errors from the other process, so that we can log
useful error messages in the requester.