feat: functions to relate job inputs & outputs (#2304)
### What
This PR enhances the dashboard's `ModelAPI` by introducing three new API
methods:
- `/api/v1/job/{id}/inputs`: Given a jobID, returns all jobs producing
input that the specified job consumes.
- `/api/v1/job/{id}/outputs`: Given a jobID, returns all jobs operating
on the output of the specified job.
- `/api/v1/cid/{cid}/jobs`: Given a CID, returns all jobs producing or
operating on the content associated with that CID.

These API methods query the underlying PostgreSQL database to find jobs
using IPFS as input or output. As a result, these queries will only work
for jobs involving data stored on or retrieved from IPFS. The queries
target the job table and examine the JSON data within the jobdata and
statedata columns. These queries may be affected by changes to the Job
Spec, so they should be maintained accordingly as the spec evolves.

Additionally, this PR updates the dashboard's job frontend to display
the results of the new API methods for the job being viewed. This allows
users to inspect related jobs and navigate between them within the
dashboard. A screenshot is provided below to illustrate these
changes:![JobRelationView](https://user-images.githubusercontent.com/6546409/231025611-656093dd-1b8c-4397-804f-13346314a029.jpg)


The `/api/v1/cid/{cid}/jobs` endpoint is currently not used by the
dashboard as I lack the required frontend skills to make that change.
However the api is callable via:
`/api/v1/cid/QmSCWyNhhUhFPQM8WoAhCt7mhJt4iN5SbpGkkS8Rktt6QV/jobs` for
example. @binocarlos if you are able to help with the frontend work here
that would be great.

---------

Co-authored-by: frrist <forrest@protocol.ai>
Co-authored-by: Walid Baruni <wdbaruni@gmail.com>