Make log-follower wait for container to exist (#2380)
It's possible to follow the logs on an execution that has not actually
started running yet (it's pre-processing things), and so attempts to
follow the log fail in unfortunate ways.

This PR introduces a flag (channel) that is used to mark, within the
docker executor, that the container actually exists. Any requests to
follow the logs will block waiting for the container to be created.

An issue remains, if the container is created and immediately causes an
error, it can be at the end of it's lifetime (and cleaning up) leaving
the log streaming trying to read from a non-existent source when
following a 'docker run'. In these cases running the logs command will
show the output (with the usual caveat about completed tasks).