Logging for wasm (#2301)
Adds support for logging to the webassembly executor, by providing a Log
Manager structure to handle interactions with logs.

When a new Wasm execution is started, the newly created log manager will
provide 2 writers to the wasm runtime, which both write to a single
ringbuffer, from which log lines are dequeued to a json file (in a
similar structure to docker's jsonfile logs). At the end of the
execution, we provide readers (one for stdout, one for stderr) which can
be used for reading the previously written logs.

Any request wanting to follow the logs will read from the log file,
switching to reading from the broadcasting goroutine once the file
contents are consumed.