Enable download of single file from output (#2345)
When fetching output, the entire folder of content is retrieved. This is
because the CID we have available is an IPFS 'directory' with child
items, and when we retrieve it the IPFS library takes care of
downloading the children to a local folder.

After an execution is complete, this is the only source of `stdout` and
`stderr`, as the compute node will have removed the working logs. Both
of these files are required by the logs command, to be able to provide
stdout and stderr content to the use. It is necessary to be able to
download individual files from the results to avoid downloading
potentially large files that will not be needed.

This PR adds a mechanism to download a single file from the output, by
adding a new DescribeResult method to the Downloader trait, providing
CIDs for each item in the IPFS directory (the base CID we have). This is
also added to the download options allowing for `bacalhau get --only
stdout ccd45d7d` and as download is handled by the client directly this
is available to other commands that download the results.