Node operators can disable features with CLI flags
There are now a set of `--disabled-*` flags to cover engines, storage,
publishers and verifiers available for the `serve` and `devstack`
commands. When used, these flags cause the specified types to be
unavailable to the whole system, even if they are otherwise installed.

This allows, for example, an operator to prevent Docker jobs from
running even if Docker is installed and accessible.