Show Client ID in `bacalhau id` output
The client id is currently useful as a persistent ID associated with
every client, and is used for things like authorisation and stats
tracking. Previously, the only way to learn one's client ID was to
submit a job to the network and get it back with `bacalhau describe`.
Now, the client id is included in output from `bacalhau id`.