Ensure all nodes have announced before running devstack tests
Since we switched to async node advertisement, our tests have been
very flaky because sometimes the node advertisement occurs after the
job has been submitted, and we have no synchronisation around waiting
for it to have happened.

Now, we ensure that we have complete information for all the nodes in
the stack before proceeding with the test.

This also makes clear when the ComputeNodeInfo has been provided or
not, which allows ranking strategies to better distinguish between
nodes that don't support anything and nodes that haven't announced.