Use the docker parsing code to parse image names
Switches to using the docker distribution Reference code for parsing the
image IDs, but stopped short of introducing the type more widely because
of complexity in it's use and excessive type checking.

Also introduces the platform information into the details returns from
the (newly renamed) ImageDistribution function which can also serve for
the ImagePlatforms call.