Only show moderation panel when there is a request
This means that we if we want to support Datacap moderation again in
the future, we will also need to populate the database with Datacap
moderation requests.

This does mean that it is now possible to have more than one request
of the same type, which will be useful when we introduce result
moderation.