Use content-disposition for downloads where available
Occassionally we get strange behavious where the filename during a
download is set to a strange value, because the original URL redirects
to a CDN which uses a different name. In these cases the response often
contains a content-disposition telling us the intended filename.

This PR reads the content-disposition and uses it for the filename when
it is available.