Verify executions using an external web hook.
This commit adds an external verifier that publishes the results and
then calls out to an external webhook with details of job executions.
The webhook can then respond with a yay or nay for each execution
individually.

There is an obvious TODO here, which is that in this scheme the
results are persisted unencrypted and visible. So if the client has
access to the storage location, they can just read the results before
the verification is done. There are some obvious solutions to this,
such as passing the data encrypted or allowing the compute provider to
configure a separate storage location for publication instead. This
will be revisited in a later commit.

There is another issue, which is that verification has until now been
something done to help the *user*, not help the compute provider. This
is fine, but given we are introducing the external verifier to allow
result moderation, this is then working against this use case.
Practically it's weird that the user has to specify
`--verifier=external` when this is not something that they should have
any control over. We could address this in a later commit by, e.g.,
allowing compute providers to specify "required verification" or by
separating these two stages completely from each other.