Run executors in an 'environment'
Currently executors have reponsibility for executing 'code', but also
has responsibility for preparing the environment where it runs. This
leads to a situation where each of the executors performs it's own
retrieval from storage and it's own mapping onto whatever mechanism it
uses to make the data available.  This makes it difficult to have a
single location where we can ensure we clean up that environment as
StorageProviders are attached directly to the executor.

This commit introduces an `executor.Environment` (cyclic dependencies
made `compute.Environment` untenable) which can prepare the storage
before providing enough information to the executor to map the data into
it's compute-space. This new environment can also ensure that any
resources created by the execution can be cleaned up once the process
has completed.

At the moment the environment still relies on the executor having a
local StorageProvider until a way can be found for the environments to
have responsibility for managing them.  Once this is achieved, there
will be need for the `executor.Executor` interface to provide
`GetStorageProvider` and this should also simplify tests.

TODO: Need to find an effective way to provide the NoopStorageProvider
to the NoopExecutor.