Use correct hash when querying locally
When querying the local dockerd to ask if it knows about an image, we
were previously using the ID which is the hash of the image itself, not
the manifest.  This commit now changes it to pull the digest from the
RepoDigests field of the image, carrying on with the remote query if not
available.