Storage cleanup of inputs (#2449)
Cleanup downloads after the executor's Run method has finished.

Currently each executor is responsible for performing the download of
content and then mapping it to their equivalent of docker's volumes.
Indirectly this has led to multiple ways of implementing it within the
executors, and triggering the cleanup task is not straight-forward.

This PR ensures it defers a function call so that the downloaded inputs
are cleaned up in line with the implementation within the storage
provider (e.g. some are noops). This implementation also now duplicates
almost identifical code but this is resolved in #2466 which is a longer
term solution to the problem of cleaning up resources.