Resolve docker digests from supplied image ids (#2410)
Updates the requester to allow it (if docker is installed) to obtain the
digest for a given docker image and use it in place of the user-provided
image. It does this by translating a docker image identifier
(repo/org/name:tag) to a docker image digest (repo/org/name@hash). As
this requires an API call to the docker instance running locally (which
may then call to a remote repository) the map from the original
identifier to the digest version is maintained in a small local cache.

Currently the caching rules for how long the map from identifier->digest
should live follows those described in
https://github.com/bacalhau-project/bacalhau/issues/2254, but we may
want to reduce the the no-tag/latest cache to avoid caching no-tag
images too long if the user is in a dev-test-push cycle of getting the
image up to date. This means that once #2254 is implemented it will only
be caching images based on their digests.

The manifest cached here is the non-platform specific manifest, e.g. for
the current ubuntu:latest it is
`ubuntu@sha256:dfd64a3b4296d8c9b62aa3309984f8620b98d87e47492599ee20739e8eb54fbf`.
This is the value that we will pass through to the compute node.
Resolving that identifier will give
```
$ docker manifest inspect ubuntu@sha256:dfd64a3b4296d8c9b62aa3309984f8620b98d87e47492599ee20739e8eb54fbf

{
   "schemaVersion": 2,
   "mediaType": "application/vnd.oci.image.index.v1+json",
   "manifests": [
      {
         "mediaType": "application/vnd.oci.image.manifest.v1+json",
         "size": 424,
         "digest": "sha256:ca5534a51dd04bbcebe9b23ba05f389466cf0c190f1f8f182d7eea92a9671d00",
         "platform": {
            "architecture": "amd64",
            "os": "linux"
         }
      },
      {
         "mediaType": "application/vnd.oci.image.manifest.v1+json",
         "size": 424,
         "digest": "sha256:2faed463fb00a57a51cc1fe0e0884d46eacac8e7784ca7a93c3e861661d3e752",
         "platform": {
            "architecture": "arm",
            "os": "linux",
            "variant": "v7"
         }
      },
...
```

The compute node will then have to look up the digest for it's chosen
platform, which in the example above, is
`sha256:ca5534a51dd04bbcebe9b23ba05f389466cf0c190f1f8f182d7eea92a9671d00`
for linux/amd64 which will return the following. This is what the
compute node should cache (the platform specific manifest).

```
$ docker manifest inspect ubuntu@sha256:ca5534a51dd04bbcebe9b23ba05f389466cf0c190f1f8f182d7eea92a9671d00
{
"schemaVersion": 2,
"mediaType": "application/vnd.oci.image.manifest.v1+json",
"config": {
    "mediaType": "application/vnd.oci.image.config.v1+json",
    "size": 2299,
    "digest": "sha256:3b418d7b466ac6275a6bfcb0c86fbe4422ff6ea0af444a294f82d3bf5173ce74"
},
"layers": [
    {
        "mediaType": "application/vnd.oci.image.layer.v1.tar+gzip",
        "size": 29534634,
        "digest": "sha256:dbf6a9befcdeecbb8813406afbd62ce81394e3869d84599f19f941aa5c74f3d1"
    }
]
}
```

The docker cli will however accept either of these tags when passed to
`docker pull` and it will do the correct thing, but it isn't clear if
both required calls (from non-specific -> platform-specific) count
towards the rate-limit.

This should close #2255