Cache supported docker image platforms in compute
The compute node currently has to query the docker hub every time it
wants to bid on a job (unless the image is already local) and this uses
our rate limit.  Instead we will cache the details for a day so that the
compute node knows whether the image is available on its platform
without the remote API call.