Record Canary Jobs in Postgres Database
We want to be recording canary jobs in the database and dashboard, with the option to filter them out further down the line, so removed canary filters, but ensured that their conditions would appear in the database so that they can be indentified at a later date. Only affects new canary jobs.

Co-authored-by: Simon Worthington <simonwo@users.noreply.github.com>