Cache docker manifests for compute node bidding (#2491)
During bidding for jobs, each compute node queries the local store, and
then the remote docker hub API to determine if the image is available
for it's platform. This happens frequently and uses our rate limit, and
so ideally we would not need to request the details for each image very
frequently.

This PR adds a small cache for caching this information, meaning that
each compute node should only need to query each image digest once ever
24 hours or so, with the result being held in RAM until it is required
in the next bid.