Exploratory Sqlite/GORM based object store
An exploration on how we can provide an object store using sqlite/gorm
as the persistence/query layers.  Even without taking advantage of the
relational nature of sqlite we encounter problems that make it difficult
to meet the brief for our object store.

* If we attempt to build a kv store in sqlite, then we're not really
  taking advantage of sqlite's best features.

* It's necessary to build DTOs (Data Transfer Objects) for every entity
  we want to store in the database to provide the timestamps. The
  alternative is to mark up our current structures with orm specific
  tags and embedded an ORM provided structure.  This is likely to get
  unwieldy very quickly.

* It's hard to provide a deep module with a narrow interface, instead
  the use of the ORM leads towards a very wide interface that is
  reasonably shallow with us barely obscuring the use of sqlite/gorm.

* Working with GORM is error prone as it requires pointers everywhere
  when `interface {}` is used in every function, so passing `Product{}`
  instead of `&Product{}` into a function will compile and panic at
  runtime.

It might be possible to build a Repository interface on top of the
sqlite/gorm so that it could be bound to each type to improve the
interface, but by the time we implement this on top of an ORM on top of
Sqlite it feels like we've strayed quite a way from the simplicity of an
object stores typical Put/Get interface.