Pin the curl docker image used in tests (#2558)
We currently don't specify a tag, and we recently got a broken image.
This pins us to 8.1.0 which we know works.

In addition we have changed the CI call from `go get` to `go mod tidy`
as `go get` seems to discover extra packages that we don't actually use,
causing the go mod check to fail. By doing this replacement, and
according to the docs below we end up with the same result (a modified
go.mod/go.sum if it is out of date).

```
Tidy makes sure go.mod matches the source code in the module.
It adds any missing modules necessary to build the current module's
packages and dependencies, and it removes unused modules that
don't provide any relevant packages. It also adds any missing entries
to go.sum and removes any unnecessary ones.
```